Move json handling to own module
